Moved testing logic into map-overlays pipe
